revamp code symbology, indices
